@@ -15,16 +15,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.4",
-    "@angular/common": "5.2.4",
-    "@angular/compiler": "5.2.4",
-    "@angular/core": "5.2.4",
-    "@angular/forms": "5.2.4",
-    "@angular/http": "5.2.4",
-    "@angular/platform-browser": "5.2.4",
-    "@angular/platform-browser-dynamic": "5.2.4",
-    "@angular/router": "5.2.4",
-    "@angular/service-worker": "5.2.4",
+    "@angular/animations": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@angular/service-worker": "6.0.3",
     "assert": "^1.4.1",
     "body-parser": "^1.18.2",
     "chart.js": "^2.7.1",
@@ -40,18 +40,18 @@
     "morgan": "^1.9.0",
     "ng2-charts": "^1.6.0",
     "path": "^0.12.7",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.2.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "5.2.4",
-    "@angular/language-service": "5.2.4",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/moment": "^2.13.0",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -63,6 +63,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
   }
 }
