@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 
 import { HttpClient } from '@angular/common/http';
-import {Observable} from "rxjs/Observable";
+import { Observable } from 'rxjs';
 
 import * as moment from 'moment';
 import * as _find from 'lodash/find';
