@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 
 import {UserService} from "./user.service";
