@@ -2,9 +2,9 @@ import { MongoClient } from 'mongodb';
 
 import {ObjectID} from "bson";
 
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
-import {Subject} from "rxjs/Subject";
+import { Observable } from 'rxjs';
+import { ReplaySubject } from 'rxjs';
+import {Subject} from 'rxjs';
 
 import * as _ from 'lodash';
 
